More comments on code
minor fix - in Drive.driveValues, also invert turnRate with reverse so
we don’t get turn inversion.  Remove the “fix” from JoystickDrive that
was inverting it there.

+++ Robot_2016/src/org/usfirst/frc3219/Robot_2016/subsystems/FeedMech.java
@@ -34,7 +34,7 @@ public class FeedMech extends Subsystem { //This class is for Putting the ball i
 		return feederLimitSwitch.get();
 	}
 	
-	public void spinFeeder(int direction) {
+	public void spinFeeder(double direction) {
 		feederMotor.setInverted(true);
 		feederMotor.set(0.3 * direction);
 	}
